Improve reliability of unapplying speed adjustment on exiting editor
Probably closes https://github.com/ppy/osu/issues/33230.

I say "probably" because I couldn't reproduce this myself using the
scenario provided in the issue but looking at the code involved I can
see why it would happen. Long and short of it is that the speed
adjustment cleanup code was much too reliant on disposal executing
quickly, which as we've learned on several occasions before, cannot be
relied upon.