Fix scroll layout height not being set early enough
Because of this it was possible for a scroll to not be able to scroll
to the panel it wanted to scroll to because its layout height was
too small for the target Y position of the panel that it wanted to
scroll to.

The scroll height was set in `updateDisplayedRange()` which is notably
after attempting to scroll to the current selection.