Disallow selected mods from being valid for freestyle as required mods due to them not being consistently compatible with other mods across rulesets
Closes https://github.com/ppy/osu/issues/33444.

The issue here is that for a set of required mods to make sense in the
context of a freestyle playlist item, it must be consistently either
valid or invalid *as a combination* across all four default rulesets,
because freestyle also permits changing ruleset. There are two
pertinent cases here:

- Flashlight and Hidden are compatible in osu!, taiko, and catch, but
  not compatible in mania. In this case I've disallowed both mods
  because of symmetry, basically - I don't see one "better mod" to
  disallow here.

- Accuracy Challenge and Easy are incompatible in osu!, catch, and mania
  (because the mod gives extra lives) there, but *compatible* in taiko,
  where it does not. In this case I've disallowed Accuracy Challenge
  only, because I find its value in being forced on a freestyle room
  to be much smaller than Easy's.

In the large scale of things I don't see this being very important
because my view is that 99% of the use case of required mods in
freestyle is going to be changing the track speed. So I don't think
anyone is going to care about this going away - but we can reassess if
I'm proven wrong.