Change lazer's valid filename method to match stable
On revisiting the issue at hand, this honestly seems like the best way
forward. It also addresses my concerns that with the method we were
using, filenames could end up being half underscores.

The main reason for choosing to change the lazer end is that stable's
difficulty update process is based on sending the beatmap's filename to
the server. This means that if we use a proposed fix of checking online
ID, it will still mean beatmaps cannot be updated on stable (for all
users which have downloaded the beatmap) if a mapper updates once on
lazer.

Implementation inspired by:

- https://referencesource.microsoft.com/#mscorlib/system/io/path.cs,1144ad3c4eff3f24
- https://github.com/peppy/osu-stable-reference/blob/996648fba06baf4e7d2e0b248959399444017895/osu!/GameplayElements/Beatmaps/Beatmap.cs#L1575-L1590

Closes #33060.