Fix lack of slider encode-decode stability due to truncating control point coordinates
Mostly closes https://github.com/ppy/osu/issues/33505.

Compare

	https://github.com/ppy/osu/blob/e82dbbfb0ed1bd5b529c0f38f72943ed32170958/osu.Game/Rulesets/Objects/Legacy/ConvertHitObjectParser.cs#L56-L59

I say "mostly" here because I'm rather skeptical that this is 100% rock
solid still, for one reason - namely that the game stores path control
point coordinates relative to the head, then turns them into absolute
coordinates when encoding, and then on decoding turns them back into
coordinates relative to the head, which in floating-point world is a Bad
Idea because of round-off error. But I'm not fixing that without
introducing a completely new beatmap format or rewriting half the
editor to address that, so I'll just pretend that I don't know any
of this until someone notices.