Restructure `CurrentlyOnlineDisplay` like `FriendsDisplay`
Using `FriendsDisplay` as a reference - that is - the "display"
component is single-instance whereas the "list" component is recreated
with a change in panel type.

I've kept the async load part in the display component, though it
doesn't have much use right now. I intend to provide the list component
an initial set of users to populate with that will handle the majority
of the load process, but that can come later.