Delay initial panel loads to boost performance
I tried two other approaches here:

- Using BDL to load the initial set of panels. This leads to a 10+
second initial load as we query 1000 users 50 at a time. We really
shouldn't be slapping the API like this, but that will require a bit
more work to get past.
- Staggering loads so panels load 5ms apart. This works but looks very
ugly.

But, this is the best I've come up with, both from a visual and lag
standpoint.

There are many other things that can be done here, including pooling
(probably complex to add in) and `VirtualisedListContainer` (doesn't
really work because that has an internal `ScrollContainer` whereas this
is being scrolled by a parent).

This is just an initial step without complicating things too much.