Add block / unblock option to user profile overlay
This is not doing the thing that the website does wherein the entire
user profile is replaced by the message that the user is blocked if
they're blocked. Someone else can try doing that.

I'm also not adding report button to this because it's going to be
annoying to make happen because currently reporting is only available as
a popover and not as a dialog. Someone else can pick that up as well.