SongSelectV2: Fix random selection not working as expected when displaying individual difficulties
Previously, random selection would always be done at a *set* level. The
final operation of a random action would be "select the user's
recommended difficulty from this randomly selected set".

This makes no sense when sets are not grouped together at song select.
In fact, it is completely broken with the previous commit which adds
group-isolated random support – if we're grouping by difficulty and the
user's recommendation is not in the current group it would throw the
user into another group unexpectedly.

This fixes the issue by splitting out the random implementation into two
separate pathways depending on the carousel display mode.