Remove incorrectly placed `ensureGlobalBeatmapValid` call
This is being run in the flow where we are providing a specific beatmap
for immediately selection. In an edge case scenario, the carousel may be
pending on a filter operation, which would cause the whole `SelectAndRun`
call to fail when it doesn't need to.

This is reproduced by multiple test scenes. One example is
`TestSceneOpenEditorTimestamp.TestErrorNotifications`.