Fix user tags moving in the control after voting
Closes https://github.com/ppy/osu/issues/33877.

Most likely regressed when the user tags were changed such that the
loading spinner that shows on adding/removing a vote was introdiced to
every individual tag separately. This in turn means that the
`LoadingLayer` responsible for showing the spinner also briefly consumes
all input when visible, which also means that the control briefly
becomes unhovered, breaking the logic.

This probably doesn't work on mobile because mobile input sucks. On iOS
simulator it looks somewhat fine in that the tags don't move until you
touch the screen anywhere else which seems okay if that's what actually
what happens on device as well. And if it isn't I'm not sure I can do
anything sane about it anyway.