Revert blocking resume input from reaching slider head circles
It doesn't work because `SliderInputManager` is too complex.

While the resume overlay *can* block the initial press, later things
break down because of

	https://github.com/ppy/osu/blob/e713a68c4936b7f4d9f84d2904dd5e995f7b8820/osu.Game.Rulesets.Osu/Objects/Drawables/SliderInputManager.cs#L255

through which the `SliderInputManager` manages to observe the keypress
it was not supposed to ever see (because the resume overlay was supposed
to be handling it).

I'm not making the `SliderInputManager` logic even more complex to
attempt to bypass this. Even now I'm not sure how I would ever make this
work, that class is way too difficult to make sense of as is anyway.