Remove song select background blur setting
Was quite unpopular amongst users. Rather than reimplementing in song
select v2, let's just remove the setting and default to never blurring
the background at song select.

We may add support for doing this back if there's enough people that
want it, but it will likely be exposed in the skin editor rather than as
a global setting like this.