Fix skins containing subdirectories breaking on external edit on windows
Closes https://github.com/ppy/osu/issues/33994.

The reason for the breakage is that `Directory.EnumerateFiles()` used in

	https://github.com/ppy/osu/blob/b1435d35e56eed08a57d1909fa0b16e67bd9c2a2/osu.Game/Skinning/SkinImporter.cs#L63

will use the primary platform directory separator character, which is
`\` on windows and `/` on unices. The internal realm storage structure
is expecting paths to be normalised to the unix convention, which is
evident in

	https://github.com/ppy/osu/blob/b1435d35e56eed08a57d1909fa0b16e67bd9c2a2/osu.Game/Database/RealmArchiveModelImporter.cs#L499

on the write side and in

	https://github.com/ppy/osu/blob/b1435d35e56eed08a57d1909fa0b16e67bd9c2a2/osu.Game/Skinning/RealmBackedResourceStore.cs#L50

on the read side.

Rather than applying this locally to the skin importer I kinda think
it's better to have this call in `ModelManager` to hopefully avoid
future footgunnage of this kind.