Fix editor metadata section reload potentially confusing title and author
Probably closes https://github.com/ppy/osu/issues/34449. I say probably
because I couldn't break this myself without commenting out chunks of
code that apply the metadata from the id3 tags, and the reporter of that
issue did not provide mp3s to test with or a clear reproduction
scenario. But chances are this is going to fix the problem anyhow.