Fix clicking group & set headers not working (or crashing) during a filter
This papers over most of https://github.com/ppy/osu/issues/34507.

The full failure scenario is:

- user selects a beatmap set
- online lookups fire
- online lookups are populated back to realm
- the realm changes trigger subscription, which triggers detached store
  replace, which triggers refilter
- refilter takes a while
- during the refilter user may be unable to select some sets because the
  set header mapping is incomplete

This only papers over the last bullet of that, in making sure that the
set header mapping is not externally accessed while it's being rebuilt.
It doesn't fix the insane amount of refilters caused by everything
preceding the refilter being overeager to trigger said refilter.