Be more aggressive about cancelling grouping filter
This is also papering over the larger issue of the insane refilter count
and is basically a drive-by fix but rider is bugging me with a bunch of
yellow highlights from memory churn which looks somewhat easily
preventable.

Basically the cancellation of filters could only take place between full
groups. If there are few groups with lots of beatmaps, it could be a
*looooong* while before the cancellation of the filter would actually
take place. Hence the added extra cancellation check before every
individual group item, and one extra check before performing the set
mapping exchange and items return.