Fix `FilterCriteria` only being able to support one tag filter at once
Other textual keyword filters also worked like that, wherein if you did
`artist=a artist=b` the second filter would overwrite the second, but in
those cases the query is against a single field, so attempting to put
multiple search criteria in conjunction on a single field is kind of
nonsensical, so it was sort of fine to do that.

Which is not the case for user tags, which are multi-valued.