Ignore more potentially incorrect data from BASS
This is still a workaround but arguably it's something we could leave in
place without much loss. I think this at least feels better than the
previous code.

Notably, you could argue the test coverage of the fail case is lower
since made it implicit that all tests will avoid the "backwards seek"
detections.  But we never really had tests correctly- fail on the original
so I don't see any loss of value.