Remove failing test
Test was added in 5495c2090a76bfd6f6a5efd0fd61530395cc6a68 to exercise a
previous version of the hotfix-workaround of not allowing seeking back.
Now that the hotfix-workaround also encompasses seeking *forward*, which
breaks another 100 tests, thus requiring the hotfix-workaround to be
turned off for tests, this situation is just completely out of
control. So I'm just removing the test because ugh.