Remove failing test
Test was added in 59edf33a5d9d055da7b01ef6838fd028815ae58c to exercise a
previous version of the hotfix-workaround of not allowing seeking back.
Now that the hotfix-workaround also encompasses seeking *forward*, which
breaks another 100 tests, thus requiring the hotfix-workaround to be
turned off for tests, this situation is just completely out of
control. So I'm just removing the test because ugh.