Attempt to fix test failures
That `SongSelect` change is a bit of a hail mary. A bunch of tests are
failing because of "DummyAPIAccess cannot process this request" and I
cannot either replicate or really understand why. The only thing that
makes some semblance of sense to me is that this is unobserved
exceptions bubbling or something because potentially `Exception` is not
accessed if the scheduled delegate never runs.