Fix broken reselection logic
fd04f7246113fd9c67ffcf184a05bb9dbc895945 contains a sneaky regression.
The previous code read:

    if (CurrentSelection != null && CheckModelEquality(beatmap, CurrentSelection))
	RequestSelection(matchingNewBeatmap);

and the new one reads:

    if (CurrentSelection is GroupedBeatmap currentBeatmapUnderGrouping)
    {
	var candidateSelection = currentBeatmapUnderGrouping with { Beatmap = beatmap };
	if (CheckModelEquality(candidateSelection, CurrentSelection))
	    RequestSelection(candidateSelection);
    }

The point is that we want to reselect `matchingNewBeatmap` here, not the
old selection. The `CheckModelEquality()` check's purpose is to check
whether *the current selection needs updating*.

I'm not sure why tests just wonderfully passed despite this, but my
suspicion is that it was because of accidental copying of realm guids
that obscured this problem.