Adjust song select debounce to be high than `repeat_initial_delay`
I'm doing this silently to see if any users complain without being told
about the change. Without this, when holding left arrow for short
bursts, precisely *one* beatmap change happens before actual key repeat
kicks in, which feels really weird (updates the leaderboard / background
unexpectedly).

This is the simplest way to resolve the issue, so if users aren't
offended by it I think we should commit to it.

Personally it's still fast enough to not annoy me at all.