Disable hitobject dimming unless peppy
I have been informed this is not a "core gameplay mechanics" and no one
cares about it. As the founder of this game I disagree, but if I'm the
only one then this seems like an amicable path forward.

If you also care, please upvote this issue and I will consider expanding
the inclusion list of user which want dimming. But do note that there's
a rank cutoff for this because apparently *this* is the breaking point
for what is considered fair vs non-fair game-localised gamma adjustment.