Fix song select not changing global beatmap correctly when switching rulesets
Closes https://github.com/ppy/osu/issues/35113.

Regressed in dfed564bda7408ec4fd1c82ab15f90f410cdd444 - setting
`Carousel.CurrentSelection` was not all that
`requestRecommendedSelection()` was doing there...

A potential point of discussion is whether this global beatmap switch
should be debounced or instant. I'm not sure I have a particularly
well-formed opinion on that. One argument in favour of not debouncing is
that if you look closely at the left side of the screen while the
debounce is in progress, you can still sort of see the broken behaviour
happen - it just doesn't stay there forever.

Thankfully `ensureGlobalBeatmapValid()` being called in every scenario
on screen suspension prevented this bug from being any worse than it is
right now.